+++ src/stores/favorites.js
@@ -1,47 +1,102 @@
-// stores/favorite.js
-import { defineStore } from 'pinia'
-import axios from 'axios'
+import { ref } from "vue";
+import axios from "axios";
+import { LoginModalStore } from '@/stores/LoginModal.js'
 
-export const useFavoriteStore = defineStore('favorite', {
-  state: () => ({
-    favorites: []
-  }),
-  actions: {
-    // 新增收藏
-    async addFavorite(placeId) {
-      try {
-        const response = await axios.post('/api/favorites', { placeId })
-        this.favorites.push(response.data)
-        return response.data
-      } catch (error) {
-        console.error('收藏失敗', error)
-        throw error
-      }
-    },
+const LoginStore = LoginModalStore()
+const API_URL = "http://localhost:3000";
+const favorites = ref([]); 
+const userId = ref(localStorage.getItem("userId"));
+const token = localStorage.getItem("token"); 
 
-    // 取得收藏列表
-    async fetchFavorites() {
-      try {
-        const response = await axios.get('/api/favorites')
-        this.favorites = response.data
-        return this.favorites
-      } catch (error) {
-        console.error('取得收藏列表失敗', error)
-        throw error
-      }
-    },
+// 檢查是否已收藏
+const isFavorited = (placeId) => {
+  return favorites.value.some((fav) => fav.favorite_places === placeId);
+};
 
-    // 移除收藏
-    async removeFavorite(placeId) {
-      try {
-        await axios.delete(`/api/favorites/${placeId}`)
-        this.favorites = this.favorites.filter(f => f.placeId !== placeId)
-      } catch (error) {
-        console.error('取消收藏失敗', error)
-        throw error
-      }
-    }
+// 載入收藏列表
+const loadFavorites = async () => {
+  if (!userId.value || !token) return;
+
+  try {
+    const response = await axios.get(`${API_URL}/favorites/${userId.value}`, {
+      headers: { Authorization: token },
+    });
+    favorites.value = response.data; // 更新收藏列表
+  } catch (error) {
+    console.error("無法加載收藏資料:", error);
   }
-})
+};
+
+// 切換收藏狀態
+const toggleFavorite = async (item) => {
+  if (!userId.value || !token) {
+    LoginStore.openModal();
+    return;
+  }
+
+  const headers = { Authorization: token }; 
+
+  try {
+    if (!isFavorited(item.id)) {
+      await addFavorite(item, headers);
+    } else {
+      await removeFavorite(item.id, headers);
+    }
+  } catch (error) {
+    console.error("操作失敗:", error);
+    alert("操作失敗，請稍後再試");
+  }
+};
+
+// 新增收藏
+const addFavorite = async (item, headers) => {
+  const placeData = {
+    place_id: item.id,
+    name: item.name,
+    image_url: item.url,
+    address: item.location,
+    rating: item.rating,
+    google_map_url: item.mapUrl,
+  };
+
+  // 新增地點
+  await axios.post(`${API_URL}/places`, placeData, { headers });
+
+  const favoriteData = {
+    favorite_user: Number(userId.value),
+    favorite_places: placeData.place_id,
+  };
+
+  // 新增收藏
+  const favoriteResponse = await axios.post(
+    `${API_URL}/favorites`,
+    favoriteData,
+    { headers }
+  );
+
+  // 更新收藏列表
+  favorites.value.push({
+    favorite_places: placeData.place_id,
+    ...favoriteResponse.data,
+  });
+};
+
+// 移除收藏
+const removeFavorite = async (placeId, headers) => {
+  await axios.delete(`${API_URL}/favorites`, {
+    data: {
+      favorite_user: Number(userId.value),
+      favorite_places: placeId,
+    },
+    headers,
+  });
+
+  // 更新收藏列表
+  favorites.value = favorites.value.filter(
+    (fav) => fav.favorite_places !== placeId
+  );
+};
 
 
+
+export { favorites, isFavorited, loadFavorites, toggleFavorite };
